+++ node-api/handlers/beacon/genesis_test.go
@@ -26,11 +26,11 @@ import (
 
 	"github.com/berachain/beacon-kit/log"
 	"github.com/berachain/beacon-kit/log/noop"
-	beaconecho "github.com/berachain/beacon-kit/node-api/engines/echo"
 	"github.com/berachain/beacon-kit/node-api/handlers/beacon"
 	"github.com/berachain/beacon-kit/node-api/handlers/beacon/mocks"
 	beacontypes "github.com/berachain/beacon-kit/node-api/handlers/beacon/types"
 	"github.com/berachain/beacon-kit/node-api/handlers/types"
+	"github.com/berachain/beacon-kit/node-api/middleware"
 	"github.com/berachain/beacon-kit/primitives/common"
 	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/labstack/echo/v4"
@@ -137,8 +137,8 @@ func TestGetGenesis(t *testing.T) {
 			h := beacon.NewHandler(backend)
 			h.SetLogger(noop.NewLogger[log.Logger]())
 			e := echo.New()
-			e.Validator = &beaconecho.CustomValidator{
-				Validator: beaconecho.ConstructValidator(),
+			e.Validator = &middleware.CustomValidator{
+				Validator: middleware.ConstructValidator(),
 			}
 
 			// set expectations
