+++ node-api/middleware/response.go
@@ -18,7 +18,7 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-package echo
+package middleware
 
 import (
 	"net/http"
